Move pen and wedo into extensions folder
