+++ src/extensions/scratch3_pen/index.js
@@ -1,11 +1,11 @@
-const ArgumentType = require('../extension-support/argument-type');
-const BlockType = require('../extension-support/block-type');
-const Cast = require('../util/cast');
-const Clone = require('../util/clone');
-const Color = require('../util/color');
-const MathUtil = require('../util/math-util');
-const RenderedTarget = require('../sprites/rendered-target');
-const log = require('../util/log');
+const ArgumentType = require('../../extension-support/argument-type');
+const BlockType = require('../../extension-support/block-type');
+const Cast = require('../../util/cast');
+const Clone = require('../../util/clone');
+const Color = require('../../util/color');
+const MathUtil = require('../../util/math-util');
+const RenderedTarget = require('../../sprites/rendered-target');
+const log = require('../../util/log');
 
 /**
  * Icon svg to be displayed at the left edge of each extension block, encoded as a data URI.
